requirements.txt
coverage==4.5.4
Django==2.2.7
djangorestframework==3.10.3
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
